Working on Random Name Command and added ReadMe
Added a ReadMe and started working on the first command I want this bot to possess; setting a user's Server display name to something random and preventing them from changing it back.

I have some infrastructure for the command to randomize a member's server name (and use on_member_update to prevent that person from changing it back), but I am encountering errors on both of these actions.

Perhaps this has something to do with me performing these actions on my user, which owns the bot? I'll check later.