Fixed Name Change Command
So it looks like some of the problem I was running into was the issue that I was trying to change the name of my own Discord profile (the one that owns the bot). Once I put an alt account in there and changed some code around, everything worked very nicely!

Not to say this command is done by any means. I still need to add a timeout for this command, add some method to get random semi-troll-y names, and restrict the commands this bot has to be only used by admin users in a Discord server.

Oh, I also realized that I needed the logging module for Discord.py to properly report errors/issues.