Added names, expiration time, and check for admin on randomName command
Added a list of semi-trolly names into the program, a check that prevents anyone without Administrator privilege's from running the command, and a time for the name change to expire (set to only five minutes for now).

Hope to later add the option to pass in a "Time" argument to the command, which will allow a custom time to be passed in.